+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/CheckoutPresenter.java
@@ -233,7 +233,6 @@ public class CheckoutPresenter extends MvpPresenter<CheckoutView, CheckoutProvid
     }
 
     private void showReviewAndConfirm() {
-        state.editPaymentMethodFromReviewAndConfirm = false;
         getView().showReviewAndConfirm(isUniquePaymentMethod());
     }
 
@@ -324,21 +323,11 @@ public class CheckoutPresenter extends MvpPresenter<CheckoutView, CheckoutProvid
     }
 
     public void onPaymentMethodSelectionError(final MercadoPagoError mercadoPagoError) {
-        if (!state.editPaymentMethodFromReviewAndConfirm) {
-            getView().cancelCheckout(mercadoPagoError);
-        } else {
-            state.editPaymentMethodFromReviewAndConfirm = false;
-            getView().backToReviewAndConfirm();
-        }
+        getView().cancelCheckout(mercadoPagoError);
     }
 
     public void onPaymentMethodSelectionCancel() {
-        if (!state.editPaymentMethodFromReviewAndConfirm) {
-            cancelCheckout();
-        } else {
-            state.editPaymentMethodFromReviewAndConfirm = false;
-            getView().backToReviewAndConfirm();
-        }
+        cancelCheckout();
     }
 
     public void onReviewAndConfirmCancel() {
@@ -486,7 +475,6 @@ public class CheckoutPresenter extends MvpPresenter<CheckoutView, CheckoutProvid
     }
 
     public void hook2Continue() {
-        state.editPaymentMethodFromReviewAndConfirm = false;
         showReviewAndConfirm();
     }
 
@@ -629,7 +617,6 @@ public class CheckoutPresenter extends MvpPresenter<CheckoutView, CheckoutProvid
 
     //TODO separate with better navigation when we have a proper driver.
     public void onChangePaymentMethodFromReviewAndConfirm() {
-        state.editPaymentMethodFromReviewAndConfirm = true;
         onChangePaymentMethod();
     }
 }